refactor(e2e-suite): move trading PO to its own folder
remove unused locators
