+++ suite/e2e/support/pageObjects/trading/assetsSection.ts
@@ -2,7 +2,7 @@ import { Locator, Page, expect } from '@playwright/test';
 
 import { NetworkSymbol } from '@suite-common/wallet-config';
 
-import { step } from '../common';
+import { step } from '../../common';
 
 export class AssetsSection {
     readonly section: Locator;
